+++ src/main/java/hr/tvz/serviceplanner/persistence/dao/impl/PurchaseDaoImpl.java
@@ -5,6 +5,7 @@ import org.springframework.stereotype.Repository;
 import hr.tvz.serviceplanner.persistence.dao.common.AbstractHibernateDao;
 import hr.tvz.serviceplanner.persistence.dao.interfaces.PurchaseDao;
 import hr.tvz.serviceplanner.persistence.models.Customer;
+import hr.tvz.serviceplanner.persistence.models.Group;
 import hr.tvz.serviceplanner.persistence.models.Price;
 import hr.tvz.serviceplanner.persistence.models.Product;
 import hr.tvz.serviceplanner.persistence.models.Purchase;
@@ -24,14 +25,19 @@ public class PurchaseDaoImpl extends AbstractHibernateDao<Purchase> implements P
 		Customer customer = getCurrentSession().get(Customer.class, purchase.getCustomer().getId());
 		Price price = getCurrentSession().get(Price.class, purchase.getPrice().getId());
 		Venue venue = getCurrentSession().get(Venue.class, venueId);
-
-		if (product != null && customer != null && price != null && venue != null) {
-			purchase.setProduct(product);
-			purchase.setPrice(price);
-			purchase.setCustomer(customer);
-			purchase.setVenue(venue);
-			create(purchase);
-			return purchase.getId();
+		if (product != null && product.getCategory() != null && product.getCategory().getGroup() != null
+				&& product.getCategory().getGroup().getId() != null) {
+			Group group = getCurrentSession().get(Group.class, product.getCategory().getGroup().getId());
+			if (group != null && customer != null && price != null && venue != null) {
+				if (customer.getVenue().equals(venue) && group.getVenue().equals(venue)) {
+					purchase.setProduct(product);
+					purchase.setPrice(price);
+					purchase.setCustomer(customer);
+					purchase.setGroup(group);
+					create(purchase);
+					return purchase.getId();
+				}
+			}
 		}
 		return null;
 	}
@@ -56,10 +62,6 @@ public class PurchaseDaoImpl extends AbstractHibernateDao<Purchase> implements P
 				Product product = getCurrentSession().get(Product.class, purchase.getProduct().getId());
 				originalPurchase.setProduct(product);
 			}
-			if (purchase.getVenue() != null) {
-				Product product = getCurrentSession().get(Product.class, purchase.getProduct().getId());
-				originalPurchase.setProduct(product);
-			}
 			if (purchase.getCustomer() != null && purchase.getCustomer().getId() != null) {
 				Customer customer = getCurrentSession().get(Customer.class, purchase.getCustomer().getId());
 				originalPurchase.setCustomer(customer);
@@ -77,7 +79,8 @@ public class PurchaseDaoImpl extends AbstractHibernateDao<Purchase> implements P
 	@Override
 	public boolean deletePurchase(Long venueId, Long purchaseId) {
 		Purchase purchase = findOne(purchaseId);
-		if(purchase.getVenue().getId() == venueId){
+		if (purchase.getGroup() != null && purchase.getGroup().getVenue() != null
+				&& purchase.getGroup().getVenue().getId() != null && purchase.getGroup().getVenue().getId() == venueId) {
 			deleteById(purchaseId);
 			return true;
 		}
